requirements.txt
python-dotenv==0.1.0
requests==2.21.0
plotly==3.8.1